重构结构 重构 DAO 层相关服务中，遗弃一部分代码和接口

+++ word-count-common/src/main/java/priv/dawn/wordcount/api/WordDaoService.java
@@ -15,8 +15,8 @@ public interface WordDaoService {
 
     List<Integer> saveWordCount(FileWordCountVo fileWordCountVo);
 
-    FileWordCountVo queryTopKWords(Long fileUid,Integer K);
+    FileWordCountVo queryTopKWords(Integer fileUid,Integer K);
 
-    FileWordCountVo queryWordCounts(Long fileUid, List<String> words);
+    FileWordCountVo queryWordCounts(Integer fileUid, List<String> words);
 
 }
